+++ src/plugins/variational_constraints/constraints.jl
@@ -1,0 +1,16 @@
+export MeanField, FullFactorization
+
+using TupleTools
+using StaticArrays
+using Unrolled
+using BitSetTuples
+using MetaGraphsNext
+using DataStructures
+using Memoization
+
+struct MeanField end
+
+struct FullFactorization end
+
+# include("constraints_macro.jl")
+include("constraints_engine.jl")
